requirements/local.txt
-r base.txt

psycopg2-binary==2.9.9
watchfiles==0.21.0
black==24.4.2
pytest-django==4.8.0